Simplify filter computation into utility functions, remove compat_filter from chain namespace.
