+++ builds/cmake/CMakeLists.txt
@@ -324,10 +324,8 @@ add_library( ${CANONICAL_LIB_NAME}
     "../../src/chain/block.cpp"
     "../../src/chain/chain_state.cpp"
     "../../src/chain/compact.cpp"
-    "../../src/chain/compact_filter.cpp"
     "../../src/chain/header.cpp"
     "../../src/chain/input.cpp"
-    "../../src/chain/neutrino_filter.cpp"
     "../../src/chain/output.cpp"
     "../../src/chain/output_point.cpp"
     "../../src/chain/payment_record.cpp"
@@ -469,6 +467,7 @@ add_library( ${CANONICAL_LIB_NAME}
     "../../src/utility/istream_bit_readers.cpp"
     "../../src/utility/istream_reader.cpp"
     "../../src/utility/monitor.cpp"
+    "../../src/utility/neutrino_filter.cpp"
     "../../src/utility/ostream_bit_writer.cpp"
     "../../src/utility/ostream_writer.cpp"
     "../../src/utility/png.cpp"
@@ -629,10 +628,8 @@ if (with-tests)
         "../../test/chain/block.cpp"
         "../../test/chain/chain_state.cpp"
         "../../test/chain/compact.cpp"
-        "../../test/chain/compact_filter.cpp"
         "../../test/chain/header.cpp"
         "../../test/chain/input.cpp"
-        "../../test/chain/neutrino_filter.cpp"
         "../../test/chain/output.cpp"
         "../../test/chain/output_point.cpp"
         "../../test/chain/payment_record.cpp"
@@ -723,6 +720,7 @@ if (with-tests)
         "../../test/utility/collection.cpp"
         "../../test/utility/data.cpp"
         "../../test/utility/endian.cpp"
+        "../../test/utility/neutrino_filter.cpp"
         "../../test/utility/png.cpp"
         "../../test/utility/property_tree.cpp"
         "../../test/utility/pseudo_random.cpp"
