+++ src/config/compact_filter.cpp
@@ -24,7 +24,6 @@
 #include <sstream>
 #include <string>
 #include <boost/program_options.hpp>
-#include <bitcoin/system/chain/compact_filter.hpp>
 #include <bitcoin/system/config/base16.hpp>
 
 namespace libbitcoin {
@@ -42,7 +41,7 @@ compact_filter::compact_filter(const std::string& hexcode)
     std::stringstream(hexcode) >> *this;
 }
 
-compact_filter::compact_filter(const chain::compact_filter& value)
+compact_filter::compact_filter(const message::compact_filter& value)
   : value_(value)
 {
 }
@@ -54,11 +53,11 @@ compact_filter::compact_filter(const compact_filter& other)
 
 compact_filter& compact_filter::operator=(const compact_filter& other)
 {
-    value_ = chain::compact_filter(other.value_);
+    value_ = message::compact_filter(other.value_);
     return *this;
 }
 
-compact_filter& compact_filter::operator=(chain::compact_filter&& other)
+compact_filter& compact_filter::operator=(message::compact_filter&& other)
 {
     value_ = std::move(other);
     return *this;
@@ -69,7 +68,7 @@ bool compact_filter::operator==(const compact_filter& other) const
     return value_ == other.value_;
 }
 
-compact_filter::operator const chain::compact_filter&() const
+compact_filter::operator const message::compact_filter&() const
 {
     return value_;
 }
